Affine hdf5 export (#221) (#222)
* E501.

The test file failed with message:

    'test_cuba.py:19:80: E501 line too long (80 > 79 characters)'

Fixed by reformatting.

* Added test for Affine hdf5 export.

The test will fail, if a network consisting of an Affine layer cannot be
exported. This is the case, e.g., if the layer has no shape ('none'). In that
case, h5py will interpret it as an ('o') object, which is not supported by the
hdf5 format.

* Dense -> Affine.

* Added shape.

The shape of the Affine layer is set with respect to the output of the
respective Dense layer. Previously this was 'None', resulting in an error while
exporting to hdf5 due to being interpreted as object type by h5py. Now
everything works due to being set to Float.

* Changed the shape to 'torch.Size()'.

* Added two tests for different modes of the "dynamics" flag.

Currently, both tests are failing. This is due to 'l.394' of
'/lib/dl/slayer/block/base.py' not being executed. This has to be fixed (?).

* Some corrections to the temp_file names.

* Added 'vThMant' export in 'base.py'.

* Added modiefied tests. Still failing.

* 'neuron' in ... -> 'neuron' not in. 1/2 pass.

* Set threshold=-1 and corrected 1 << 17 to 1 << 18.

* Check 'vThMant' in h5 file.

---------

Co-authored-by: PhilippPlank <32519998+PhilippPlank@users.noreply.github.com>
Co-authored-by: bamsumit <bam_sumit@hotmail.com>